Update Rust Coordinate Translator Javascript
This translator is used to turn meter coordinates into map coordinates. It assumes the map size to be 3300 meters. Inputs measured in meters can be acquired from the console commands (printpos) this will print your current position. And (world.monuments) this will print locations of generated landmarks on the map.